+++ packages/client/wallets/aa/src/api/CrossmintWalletService.test.ts
@@ -1,5 +1,5 @@
 import { CROSSMINT_STG_URL } from "../utils";
-import { CrossmintService } from "./CrossmintService";
+import { CrossmintWalletService } from "./CrossmintWalletService";
 
 jest.mock("@/blockchain", () => ({
     getApiUrlByBlockchainType: jest.fn(),
@@ -13,12 +13,12 @@ jest.mock("@/utils/error", () => ({
 }));
 
 describe("CrossmintService", () => {
-    let crossmintService: CrossmintService;
+    let crossmintService: CrossmintWalletService;
     const apiKey =
         "sk_staging_A4vDwAp4t5az6fVQMpQK6qapBnAqgpxrrD35TaFQnyKgxehNbd959uZeaHjNCadWDXrgLRAK1CxeasZjtYEq4TbFkKMBBvbQ9oinAxQf8LbHsSYW2DMzT8fBko3YGLq9t7ZiXZjmgkTioxGVUUjyLtWLeBKwNUDLgpshWjaoR7pKRnSE9SqhwjQbiK62VKiBTdA3KvHsyG9k8mLMcKrDyfXp";
 
     beforeEach(() => {
-        crossmintService = new CrossmintService(apiKey);
+        crossmintService = new CrossmintWalletService(apiKey);
     });
 
     describe("constructor", () => {
