+++ packages/client/wallets/aa/src/api/index.ts
@@ -1,1 +1,1 @@
-export * from "./CrossmintService";
+export * from "./CrossmintWalletService";
